The fields in a Panorama csv that relate to personal and socioeconomic identity

The fields in a Panorama csv that relate to personal and socioeconomic identity

Usage

DEMOGRAPHIC_FIELDS

Format

An object of class list of length 10.

See also